Выносит константы в конфиг
