+++ middlewares/authMiddlewares.js
@@ -1,4 +1,5 @@
 const jwt = require('jsonwebtoken');
+const { secretPhrase } = require('../utils/config');
 const UnauthorizedError = require('../errors/unauthorized-error');
 
 const { NODE_ENV, JWT_SECRET } = process.env;
@@ -10,7 +11,7 @@ module.exports.authMiddlewares = (req, res, next) => {
   } else {
     let payload;
     try {
-      payload = jwt.verify(token, NODE_ENV === 'production' ? JWT_SECRET : 'life-is-an-adventure,not-a-misery');
+      payload = jwt.verify(token, NODE_ENV === 'production' ? JWT_SECRET : secretPhrase);
     } catch (err) {
       next(new UnauthorizedError('Нужно пройти авторизацию'));
     }
